update file structure
git-svn-id: https://elmar.ips.cs.tu-bs.de/svn/cbp_ws1011@303 a60ba226-48e3-df11-83e3-000476a39db3

+++ optimizer/src/ExprNodes.cpp
@@ -104,11 +104,11 @@ constant* Expr_Mod::getConstant() {
 	if(typeLeft->isInteger() && typeRight->isInteger())
 		mulConst->integer = conLeft->integer % conRight->integer;
 	else if(typeLeft->isFloating() && typeRight->isFloating())
-		mulConst->floating = (int) conLeft->floating % (int) conRight->floating;
+		mulConst->floating = (float)( (int) conLeft->floating % (int) conRight->floating );
 	else if(typeLeft->isFloating() && typeRight->isInteger())
-		mulConst->floating = (int) conLeft->floating % (int) conRight->floating;
+		mulConst->floating = (float)( (int) conLeft->floating % (int) conRight->floating );
 	else
-		mulConst->floating = (int) conLeft->integer % (int) conRight->floating;
+		mulConst->floating = (float)( (int) conLeft->integer % (int) conRight->floating );
 	return mulConst;
 }
